Fixes leading spaces in date
exec_test_suite.tcl was previously using the %e
format for days, because it does not have leading zeros,
and was thought to be more ergonomic.
However, it does have leading spaces, which causes paths
to be broken apart with spaces.

Solution was to switch to %d format for days.