Fixed collect_stats time parsing bug
`expr <string>` will attempt to parse the string for some numerical
value.
However, it tries to infer the base from any leading part of the
string.
This is an issue for numbers with leading zeros, which are not
valid octal strings
because `expr` will not revert to a base 10 parse.
Fixed by using `scan %f` to force interpretation of string as
floating point number